(all): Remove autoassignment library and use RegExp to extract port info
In this commit, I remove the autoassignment library which was used to extract ports' info from file.
The autoassignment library doesn't support filepaths with Chinese charactors and it has some bugs. Also,
It will generate an AST which is useless in this project.

This commit replaces that library with 2 new files contaning several Regular Expressions.